fix: render nested route district
