+++ webpack.config.js
@@ -7,6 +7,7 @@ module.exports = {
     ],
     output: {
         path: path.resolve(__dirname, 'dist'),
+        publicPath: '/',
         filename: 'bundle.js',
     },
     module: {
